Refactor deleteTask controller to include apiError import for improved error handling

+++ src/controllers/taskControllers/deleteTask.controller.js
@@ -1,6 +1,6 @@
 import taskDeletedEmail from "../../emails/taskEmails/taskDeletedEmail.js";
 import { deleteFromCloudinary } from "../../utils/cloudinary.js";
-import { Activity, apiResponse, asyncHandler, Task } from "../allImports.js";
+import { Activity, apiError, apiResponse, asyncHandler, Task } from "../allImports.js";
 
 const deleteTask = asyncHandler(async (request, response) => {
     const {taskId} = request?.params;
@@ -16,6 +16,10 @@ const deleteTask = asyncHandler(async (request, response) => {
         .json(
             new apiResponse(404, {}, "Task not found, task may be deleted")
         )
+    }
+
+    if(foundTask.taskCreatedBy !== request.user.id){
+        throw new apiError(403, "Access denied")
     }
 
     await Task.findByIdAndDelete(taskId);
